fix issue 14 fix issue 15 fix bug in loading global orms

+++ src/test/java/org/guzz/orm/se/TestDaoRead.java
@@ -60,7 +60,7 @@ public class TestDaoRead extends DBBasedTestCase {
 			assertNotNull(user) ;
 			assertEquals(user.getUserName(), "name " + i) ;	
 			assertEquals(user.getPassword(), "psw " + i) ;	
-			assertEquals(user.getFavCount(), i) ;		
+			assertEquals(user.getFavCount().intValue(), i) ;		
 			assertEquals(user.isVip(), (i%2 == 0)) ;		
 		}
